Improve macro repeat highlighting
macro_rules! repetitions can use almost any token as the separator (any
but [, {, (, ], }, ), *, + and ?), and *, + and ? as the operator.

Before, this only supported the comma separator, and the * and +
operators.

Now, it supports any single-character operator (still incomplete, but
difficult to sanely improve further, and generally sufficient), and all
three operators. (Whitespace is still also mishandled, but meh.)